requirements.txt
blind-watermark==0.4.4
certifi==2023.5.7
charset-normalizer==3.1.0
distlib==0.3.8
et-xmlfile==1.1.0
filelock==3.13.3
greenlet==3.0.3
idna==3.4
Js2Py==0.74
msgpack==1.0.8
numpy==1.26.4
opencv-python==4.10.0.82
openpyxl==3.1.5
pandas==2.2.2
platformdirs==4.2.0
pyjsparser==2.7.1
pynvim==0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-gitlab==3.15.0
pytz==2024.1
PyWavelets==1.6.0
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.0.3
virtualenv==20.25.1
